+++ .rspec
@@ -1,2 +1,2 @@
 --color
---format=documentation
+--format=progress
