+++ lib/grape/validations/validators/default.rb
@@ -1,6 +1,6 @@
 module Grape
   module Validations
-    class DefaultValidator < Validator
+    class DefaultValidator < Base
       def initialize(attrs, options, required, scope)
         @default = options
         super
