+++ lib/grape/validations/validators/single_option.rb
@@ -1,0 +1,10 @@
+module Grape
+  module Validations
+    class SingleOptionValidator < Base
+      def initialize(attrs, options, required, scope)
+        @option = options
+        super
+      end
+    end
+  end
+end
